Fix a warning issue with ARMCC.
(based on FatPkg commit 5613b9c6a6fdf762ef26f30f22ecabc487e22eb4)

[jordan.l.justen@intel.com: Use script to relicense to 2-clause BSD]
Contributed-under: TianoCore Contribution Agreement 1.0
Acked-by: Laszlo Ersek <lersek@redhat.com>

+++ FatPkg/EnhancedFatDxe/ReadWrite.c
@@ -117,7 +117,7 @@ Returns:
   //
   // Set the position
   //
-  if (Position == -1) {
+  if (Position == (UINT64)-1) {
     Position = OFile->FileSize;
   }
   //
